Replaced mysterous bool's in opcodes with u8
I suppose, looking at the code, a bool is always represented as u8,
which is what we want anyway, but since these values are not bools but
placeholder integers that must currently always be 0, having them be
bools is kind of strange.

Replaced the reading/writing of them with VarUint7, which is maybe almost
as invalid, but works.

Resolves issue #145.