Move `slow_len` behind a `slow_assertions` #[cfg]
I've been experimenting with running `wasm-bindgen` (which uses `parity-wasm`
for decoding/encoding wasm modules) in debug mode where possible to reduce build
times on CI and otherwise make it more feasible to run locally. Unfortunately
the wasm modules that `wasm-bindgen` handles are sometimes pretty huge, notably
with lots of debuginfo and tons of items that are pruned later with `wasm-gc`
(and in general debug mode Rust binaries are quite large).

In some profilng one of the hottest functions in `parity-wasm` is the
`IndexMap::slow_len` function. It looks like this is an internal sanity check
only turned on when `debug_assertions` are enabled (which is good!). In this
commit, though, this check is moved behind a new `--cfg slow_assertions` flag
instead of the standard `debug_assertions`. Travis configuration has been
tweaked to always pass this flag, and it means that `slow_len` by default won't
get compiled into downstream crates in debug mode.

I'm curious to see what y'all think of this change! This is definitely not an
issue in release mode, so it's purely a slight optimization for debug mode. If
you'd prefer to not merge this and keep this on in debug mode by default, that's
also totally ok!