Add the "offsets" feature.
If enabled, this feature adds the offsets() function
to the Instructions struct to get the absolute offsets of every single instruction.
CodeSection has the offset() function to get the absolute offset of the code section. Both are useful for computing
the instruction offsets relative to start of the code section, for example when using the addr2line crate to
gather the original location in the source code for a given
instruction.
To achieve this, the ReadSeek trait bound was used instead of Read for all deserialize functions.
If "offsets" is enabled, ReadSeek requires Read + Seek, if not, it only requires "Read"